Added a function to unzip a zip file.
Checks if the given file is a zip file if so unzips it. The unzip content is saved in directory with the same name as the zip file its sefl. The output directory is located in the same directory as the zip file. The zip file is deleted at the end.

TODO:
- Error handling